% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{string_eukaryotes}
\alias{string_eukaryotes}
\title{DATASET_TITLE}
\format{A data frame with 476 rows and 4 variables:
\describe{
  \item{\code{taxid}}{character COLUMN_DESCRIPTION}
  \item{\code{string_name}}{character COLUMN_DESCRIPTION}
  \item{\code{new_taxid}}{character COLUMN_DESCRIPTION}
  \item{\code{ncbi_name}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
string_eukaryotes
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
