% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{timetree_newick}
\alias{timetree_newick}
\title{DATASET_TITLE}
\format{A newick file with 335 tips and taxids:
\describe{
  \item{\code{edge}}{integer COLUMN_DESCRIPTION}
  \item{\code{Nnode}}{integer COLUMN_DESCRIPTION}
  \item{\code{tip.label}}{character COLUMN_DESCRIPTION}
  \item{\code{taxid}}{character COLUMN_DESCRIPTION}
}}
\source{
\url{http://somewhere.important.com/}
}
\usage{
timetree_newick
}
\description{
DATASET_DESCRIPTION
}
\keyword{datasets}
